Rename bad package name and apply it.
