+++ extensions/odl/capabilities/src/main/java/org/mqnaas/extensions/odl/capabilities/flows/IFlowManagement.java
@@ -33,8 +33,8 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
 import org.mqnaas.core.api.ICapability;
-import org.mqnaas.extensions.odl.hellium.flowprogrammer.model.FlowConfig;
-import org.mqnaas.extensions.odl.hellium.flowprogrammer.model.FlowConfigs;
+import org.mqnaas.extensions.odl.helium.flowprogrammer.model.FlowConfig;
+import org.mqnaas.extensions.odl.helium.flowprogrammer.model.FlowConfigs;
 
 @Path("/")
 public interface IFlowManagement extends ICapability {
