+++ extensions/odl/capabilities/src/main/java/org/mqnaas/extensions/odl/capabilities/impl/ODLRootResourceProvider.java
@@ -51,16 +51,16 @@ import org.mqnaas.core.api.exceptions.CapabilityNotFoundException;
 import org.mqnaas.core.api.exceptions.ResourceNotFoundException;
 import org.mqnaas.core.impl.AttributeStore;
 import org.mqnaas.core.impl.RootResource;
-import org.mqnaas.extensions.odl.client.hellium.topology.api.IOpenDaylightTopologyNorthbound;
+import org.mqnaas.extensions.odl.client.helium.topology.api.IOpenDaylightTopologyNorthbound;
 import org.mqnaas.extensions.odl.client.switchnorthbound.ISwitchNorthboundAPI;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.Node.NodeType;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnector;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnectorProperties;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnectors;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeProperties;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.Nodes;
-import org.mqnaas.extensions.odl.hellium.topology.model.EdgeProperty;
-import org.mqnaas.extensions.odl.hellium.topology.model.Topology;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.NodeConnector;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.NodeConnectorProperties;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.NodeConnectors;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.NodeProperties;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.Nodes;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.Node.NodeType;
+import org.mqnaas.extensions.odl.helium.topology.model.EdgeProperty;
+import org.mqnaas.extensions.odl.helium.topology.model.Topology;
 import org.mqnaas.network.api.topology.link.ILinkAdministration;
 import org.mqnaas.network.api.topology.link.ILinkManagement;
 import org.mqnaas.network.api.topology.port.IPortManagement;
