+++ extensions/odl/model/src/test/java/org/mqnaas/extensions/odl/model/switchmanager/NodeConnectorsSerializationTests.java
@@ -31,10 +31,10 @@ import org.custommonkey.xmlunit.XMLAssert;
 import org.custommonkey.xmlunit.XMLUnit;
 import org.junit.Assert;
 import org.junit.Test;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.Node;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.Node.NodeType;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnectors;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.helpers.SwitchManagerModelHelper;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.Node;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.NodeConnectors;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.Node.NodeType;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.helpers.SwitchManagerModelHelper;
 import org.mqnaas.general.test.helpers.serialization.SerializationUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
