+++ extensions/odl/model/src/test/java/org/mqnaas/extensions/odl/model/switchmanager/NodesSerializationTests.java
@@ -31,8 +31,8 @@ import org.custommonkey.xmlunit.XMLAssert;
 import org.custommonkey.xmlunit.XMLUnit;
 import org.junit.Assert;
 import org.junit.Test;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.Nodes;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.helpers.SwitchManagerModelHelper;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.Nodes;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.helpers.SwitchManagerModelHelper;
 import org.mqnaas.general.test.helpers.serialization.SerializationUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
