+++ extensions/odl/model/src/test/java/org/mqnaas/extensions/odl/model/topology/TopologySerializationTest.java
@@ -34,14 +34,14 @@ import org.custommonkey.xmlunit.XMLAssert;
 import org.custommonkey.xmlunit.XMLUnit;
 import org.junit.Assert;
 import org.junit.Test;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.Node;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.Node.NodeType;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnector;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.NodeConnector.NodeConnectorType;
-import org.mqnaas.extensions.odl.hellium.switchmanager.model.PropertyValue;
-import org.mqnaas.extensions.odl.hellium.topology.model.Edge;
-import org.mqnaas.extensions.odl.hellium.topology.model.EdgeProperty;
-import org.mqnaas.extensions.odl.hellium.topology.model.Topology;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.Node;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.NodeConnector;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.PropertyValue;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.Node.NodeType;
+import org.mqnaas.extensions.odl.helium.switchmanager.model.NodeConnector.NodeConnectorType;
+import org.mqnaas.extensions.odl.helium.topology.model.Edge;
+import org.mqnaas.extensions.odl.helium.topology.model.EdgeProperty;
+import org.mqnaas.extensions.odl.helium.topology.model.Topology;
 import org.mqnaas.general.test.helpers.serialization.SerializationUtils;
 import org.xml.sax.SAXException;
